requirements.txt
PyQt5
pycryptodome